requirements.txt
email-validator>=2.0.0
dnspython>=2.4.0
